component file renamed
