+++ client/components/index.js
@@ -9,4 +9,4 @@ export {Login, Signup} from './auth-form'
 export {default as Cart} from './cart'
 export {default as Product} from './ProductList'
 export {default as Checkout} from './Checkout'
-export {default as OrderHistory} from './userOrderHistory'
+export {default as OrderHistory} from './order_history'
