+++ client/components/order_history/OrderProduct.js
@@ -1,5 +1,4 @@
 import React, {Component} from 'react'
-import {Link} from 'react-router-dom'
 
 const OrderProduct = ({product, quantity}) => {
   return (
